requirements.txt
easydict==1.10
faiss-cpu==1.7.4
Pillow==9.2.0
scikit-image==0.24.0
scikit-learn==1.2.2
tqdm==4.67.1
torchmetrics==1.7.3
xformers==0.0.20
timm==0.6.13
numpy==1.24.1
pandas==2.3.0
hdbscan==0.8.40

